feat: Improve code structure
