+++ src/AElf.CSharp.CodeOps/CSharpCodeOpsAElfModule.cs
@@ -27,7 +27,7 @@ public class CSharpCodeOpsAElfModule : AElfModule
 
         context.Services.AddTransient<IPatcher, StateWrittenSizeLimitMethodInjector>();
         context.Services.AddTransient<IPatcher, ResetFieldsMethodInjector>();
-        context.Services.AddTransient<IPatcher, ExecutionObserverInjector>();
+        context.Services.AddTransient<IPatcher, Patchers.Module.CallAndBranchCounts.Patcher>();
         context.Services.AddTransient<IPatcher, MethodCallReplacer>();
     }
 }
